refactor and move some foncionst our of main file
