+++ src/sql_funcs.rs
@@ -44,7 +44,11 @@ fn generate_struct(row: &Row, file_path: &std::path::Path) -> Result<(), std::io
 }
 
 
-pub fn add_select_funcs(rows: Vec<Row>, path: &std::path::Path, func_names: &mut Vec<String>) -> Result<(), io::Error> {
+pub fn add_basic_sql_funcs(
+    rows: Vec<Row>,
+    path: &std::path::Path,
+    func_names: &mut Vec<String>
+) -> Result<(), io::Error> {
 
 
    // re do each one to have layers and return the endpoint layer (api layer)
